requirements.txt
# Core dependencies
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
pydantic>=2.0.0

# Google Cloud dependencies
google-cloud-aiplatform>=1.36.0
google-auth>=2.22.0
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.100.0

# ADK dependencies
google-adk>=0.1.0

# Email processing
email-validator>=2.0.0

# Development dependencies
pytest>=7.0.0
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0 